Please update with London Borough of Barnet data
Hi 
I have not made any updates but I request the following: 
In your boundary map the London Borough's of Enfield and Barnet are erroneously joined as one borough and its details are 'London Borough of Enfield'.  Could you update and create a boundary for London Borough of Barnet?  

Many thanks